cmake: use standard options for shared vs. static
The built-in option BUILD_SHARED_LIBS is used to control whether
libraries should be built as shared or static.  The idea is that the
option can be set by the user for the top-level project and cascade
down to all sub-projects (e.g., including third-party libraries).

In default CMake, BUILD_SHARED_LIBS defaults to OFF and is used to
determine in unadorned 'add_library(mylib mylib.c)' invocation should
be built as shared or static.

We follow the same semantics for BUILD_SHARED_LIBS, but with an
additional BUILD_STATIC_LIBS option used to allow building both static
and shared variants.

If BUILD_SHARED_LIBS is ON, the shared variant will be built.  If
BUILD_STATIC_LIBS is ON, the static variant be built.  If both are ON,
both will be built.  If neither is ON, we force BUILD_STATIC_LIBS to
ON and build the the static library, to match the default semantics
for BUILD_SHARED_LIBS.

An optional STATIC_SUFFIX variable may be set by the user to append a
suffix (e.g., _static) to the static library name.