cmake: use standard option for disabling tests
The CMake option BUILD_TESTING is created by CTest to allow the user
to control the compilation of tests.  It defaults to ON.

Other tools, like Buildroot, set the BUILD_TESTING option for CMake
packages, so we want to support it.

Also, there is no need to manually call enable_testing() - the CTest
module does that based on the value of BUILD_TESTING.