travis: Use new ibm tpm simulator and tss.
The old version we were using appears to have some bug where the first
usage of a primary key fails, but only when ordered in some particular
way.
The old ordering of our tests didn't run into this, but the new
reorganization caused a reordering of the tests and consistently tripped
this.
The issue doesn't appear to be with our library (I was able to trigger
it using the tpm2-software TSS stack as well).

Updating to the newest IBM simulator version appears to fix it.